requirements.txt
functions-framework
requests
google-cloud-storage